+++ frontend/src/routers/Router.jsx
@@ -1,17 +1,18 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import Home from "@screens/Home";
-import Login from "@screens/Login";
+import Login from "@screens/login/Login";
 import Letters from "@screens/Letters";
-import Join from "@screens/Join";
-import JoinMail from "@screens/JoinMail";
-import JoinUserInfo from "@screens/JoinUserInfo";
-import JoinAdd from "@screens/JoinAdd";
-import JoinFinish from "@screens/JoinFinish";
+import Join from "@screens/join/Join";
+import JoinMail from "@screens/join/JoinMail";
+import JoinUserInfo from "@screens/join/JoinUserInfo";
+import JoinAdd from "@screens/join/JoinAdd";
+import JoinFinish from "@screens/join/JoinFinish";
 import MyPage from "@screens/MyPage";
 import Main from "@screens/Main";
 import StudyList from "@screens/StudyList";
-import FindPw from "@screens/FindPw";
+import FindPw from "@screens/login/FindPw";
+import NewPw from "@screens/login/NewPw";
 import StudyDetail from "@screens/StudyDetail";
 import MainNavBar from "@components/common/MainNavBar";
 import StudyRegistForm from "@screens/common/StudyRegistForm";
@@ -39,6 +40,7 @@ function Router() {
         {/* 로그인 */}
         <Route path="/login" element={<Login />} />
         <Route path="/find/pw" element={<FindPw />} />
+        <Route path="find/newpw" element={<NewPw />} />
         {/* 마이페이지 */}
         <Route path="/mypage/*" element={<MyPage />} />
