--- frontend/src/screens/JoinAdd.scss
@@ -1,93 +1,0 @@
-#join_add {
-  margin: vw(100px) vw(483px);
-
-  border-radius: vw(20px);
-  width: vw(473px);
-  height: vw(730px);
-  border: solid 1px $GRAY-DB;
-
-  .add_text {
-    margin: vw(64px) vw(70px) vw(32px) vw(72px);
-
-    &_title {
-      line-height: vw(35px);
-    }
-  }
-
-  .add_input {
-    &_title {
-      margin: 0 0 vw(8px) 0;
-      line-height: 23px;
-    }
-    input {
-      border: solid vw(1px) $GRAY-DB;
-      padding-left: vw(12px);
-      height: vw(48px);
-      width: vw(329px);
-      border-radius: vw(4px);
-    }
-
-    &_box {
-      margin-bottom: vw(16px);
-    }
-
-    &_notice {
-      color: $GRAY-76;
-      line-height: vw(17px);
-      margin-top: vw(8px);
-    }
-
-    &_email {
-      background-color: $GRAY-DB;
-    }
-
-    .join_address_btn {
-      width: vw(329px);
-      height: vw(32px);
-      background-color: $PINK-LIGHT;
-      color: $WHITE;
-      border-radius: vw(4px);
-      margin-top: vw(8px);
-    }
-  }
-  .add_category {
-    margin: vw(8px) 0;
-    text-align: left;
-    padding: 0 vw(72px) vw(12px) vw(70px);
-
-    &_title {
-      line-height: vw(23px);
-      margin: vw(8px) 0 vw(20px) 0;
-    }
-
-    &_options {
-      flex-direction: row;
-      flex-wrap: wrap;
-      justify-content: center;
-    }
-
-    button {
-      width: vw(64px);
-      height: vw(32px);
-      margin: 0 vw(12px) vw(8px) 0;
-
-      border: 1px solid #ff79ab;
-      border-radius: 10px;
-
-      &:hover,
-      &:active {
-        color: $WHITE;
-        background-color: $PINK-LIGHT;
-      }
-    }
-  }
-
-  .join_add_next_btn {
-    width: vw(329px);
-    height: vw(48px);
-    border-radius: vw(4px);
-    background-color: $PINK-MAIN;
-    color: $WHITE;
-    margin-top: vw(12px);
-  }
-}
