--- frontend/src/screens/JoinFinish.scss
@@ -1,32 +1,0 @@
-#join_finish {
-  margin: vw(150px) vw(483px);
-
-  border-radius: vw(20px);
-  width: vw(473px);
-  height: vw(720px);
-
-  .join_finish_image {
-    width: vw(120px);
-    margin-bottom: vw(39px);
-  }
-  .finish_text {
-    margin: 0 0 vw(44px) 0;
-    text-align: center;
-    &_title {
-      line-height: vw(41px);
-      margin-bottom: vw(24px);
-    }
-    &_desc {
-      line-height: vw(23px);
-    }
-  }
-
-  .finish_btn {
-    width: vw(240px);
-    height: vw(48px);
-    border-radius: vw(4px);
-    background-color: $PINK-MAIN;
-    color: $WHITE;
-    margin-top: vw(22px);
-  }
-}
