--- frontend/src/screens/JoinMail.scss
@@ -1,83 +1,0 @@
-#join_mail {
-  margin: vw(100px) vw(483px);
-
-  border-radius: vw(20px);
-  width: vw(473px);
-  height: vw(680px);
-  border: solid 1px $GRAY-DB;
-
-  .mail_text {
-    margin: vw(64px) vw(70px) vw(36px) vw(72px);
-    text-align: left;
-    &_desc {
-      line-height: vw(35px);
-      margin-bottom: vw(16px);
-      &_small {
-        line-height: vw(23px);
-      }
-    }
-  }
-
-  .mail_input {
-    // margin: 0 0 vw(16px) 0;
-    &_title {
-      margin: 0 0 vw(19px) 0;
-    }
-    input {
-      border: solid vw(1px) $GRAY-DB;
-      padding-left: vw(12px);
-      height: vw(48px);
-      width: vw(329px);
-      border-radius: vw(4px);
-    }
-
-    &_box {
-      margin-bottom: vw(12px);
-    }
-
-    &_failed {
-      color: $RED;
-      line-height: vw(17px);
-      margin: 0 0 vw(16px) 0;
-    }
-  }
-
-  .join_mail_next_btn {
-    width: vw(329px);
-    height: vw(48px);
-    border-radius: vw(4px);
-    background-color: $PINK-MAIN;
-    color: $WHITE;
-    margin-bottom: vw(24px);
-  }
-
-  .if_text {
-    color: $GRAY-76;
-    &_title {
-      margin: vw(4px) vw(8px) vw(4px) 0;
-      line-height: vw(17px);
-    }
-    margin-bottom: vw(41px);
-
-    a {
-      text-decoration: underline;
-    }
-  }
-
-  .notice_mail {
-    height: vw(160px);
-    width: vw(329px);
-    background-color: $GRAY-F3;
-    border-radius: vw(4px);
-    padding: vw(29px) vw(20px);
-    &_text {
-      color: $GRAY-61;
-      line-height: vw(17px);
-      padding-bottom: vw(17px);
-      &_em {
-        color: $GRAY-61;
-        line-height: vw(17px);
-      }
-    }
-  }
-}
