--- frontend/src/screens/Login.scss
@@ -1,100 +1,0 @@
-#login {
-  margin: vw(100px) vw(483px);
-
-  border-radius: vw(20px);
-  width: vw(473px);
-  height: vw(855px);
-  border: solid 1px $GRAY-DB;
-
-  .login_text {
-    padding: vw(64px) 0 vw(40px) vw(72px);
-    text-align: left;
-    line-height: vw(35px);
-  }
-
-  .login_social_btn {
-    button {
-      width: vw(329px);
-      height: vw(48px);
-      border-radius: vw(4px);
-      margin-bottom: vw(12px);
-
-      img {
-        margin-right: vw(8px);
-        width: vw(16px);
-      }
-    }
-    &_kakao {
-      background-color: $KAKAO-YELLOW;
-      color: $KAKAO-BROWN;
-    }
-
-    &_naver {
-      background-color: $NAVER;
-      color: $WHITE;
-    }
-
-    &_google {
-      background-color: $WHITE;
-      border: solid 1px $GRAY-CB;
-    }
-
-    margin-bottom: vw(20px);
-  }
-
-  .login_or {
-    &_line {
-      height: vw(1px);
-      background-color: $GRAY-CB;
-      width: vw(138px);
-    }
-    &_text {
-      line-height: vw(20px);
-      color: $GRAY-CB;
-      margin: 0 vw(13px) 0 vw(14px);
-    }
-  }
-
-  .login_input {
-    margin: vw(32px) 0 vw(36px) 0;
-    &_title,
-    &_failed {
-      margin: 0 0 vw(12px) vw(72px);
-    }
-    input {
-      border: solid vw(1px) $GRAY-DB;
-      padding-left: vw(12px);
-      height: vw(48px);
-      width: vw(329px);
-      border-radius: vw(4px);
-    }
-
-    &_box {
-      margin-bottom: vw(20px);
-    }
-
-    &_failed {
-      color: $RED;
-    }
-  }
-
-  .login_btn {
-    width: vw(329px);
-    height: vw(48px);
-    border-radius: vw(4px);
-    background-color: $PINK-MAIN;
-    color: $WHITE;
-    margin-bottom: vw(24px);
-  }
-
-  .if_text {
-    color: $GRAY-76;
-    &_title {
-      margin: vw(4px) vw(8px) vw(4px) 0;
-      line-height: vw(17px);
-    }
-    a {
-      text-decoration: underline;
-    }
-  }
-}
